Put `filterTree` function into own utils file
